% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_amw_data_functions.R
\name{add_concordance_codes}
\alias{add_concordance_codes}
\title{Add 3 letter ISO codes and MWTools-specific region codes}
\usage{
add_concordance_codes(
  .live_animals,
  concordance_path = MWTools::fao_concordance_path(),
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  country_name = MWTools::mw_constants$country_name,
  country_incl_col = MWTools::conc_cols$country_incl_col,
  country_code_col = MWTools::conc_cols$country_code_col,
  country_code_pfu_col = MWTools::conc_cols$country_code_pfu_col
)
}
\arguments{
\item{country_code_pfu_col}{}
}
\value{

}
\description{
Adds 3 letter ISO country codes and muscle work region codes for countries
present in the FAO's Live animals data, downloaded through the \code{FAOSTAT} package,
usually using the \code{MWTools::down_fao_live_animals} function.
}
