% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{add_hmw_analysis_sectors}
\alias{add_hmw_analysis_sectors}
\title{Map ILO broad sectors HMW analysis sector categories}
\usage{
add_hmw_analysis_sectors(
  .df,
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path(),
  sector_col = MWTools::mw_constants$sector_col,
  hmw_sector_map_sheet = MWTools::hmw_analysis_constants$hmw_sector_map_sheet,
  agriculture = MWTools::mw_sectors$agriculture_broad.sector,
  hmw_analysis_sector_col = MWTools::mw_sectors$hmw_analysis_sector_col,
  industry = MWTools::mw_sectors$industry_broad.sector,
  services = MWTools::mw_sectors$services_broad.sector
)
}
\arguments{
\item{.df}{A data frame containing the number of hours worked by broad sector
(Agriculture, Industry, and Services).
Usually produced by calling the
\code{add_hmw_region_codes},
\code{fill_ilo_data},
\code{calc_total_hours_worked}, and
\code{get_broad.sector_data} functions in sequence on the raw FAO data.}

\item{hmw_analysis_data_path}{See \code{MWTools::hmw_analysis_data_path()}.}

\item{sector_col}{See \code{MWTools::mw_constants}.}

\item{hmw_sector_map_sheet}{See \code{MWTools::hmw_analysis_constants}.}

\item{hmw_analysis_sector_col, agriculture, industry, services}{See \code{MWTools::mw_sectors}.}
}
\description{
Map ILO broad sectors HMW analysis sector categories, which correspond to
activity levels (Primary, Secondary, and Sedentary).
}
\examples{
working_hours_activity_data <- read.csv(file = MWTools::hmw_test_data_path()) \%>\%
  add_hmw_region_codes() \%>\%
  fill_ilo_data() \%>\%
  calc_total_hours_worked() \%>\%
  get_broad.sector_data() \%>\%
  add_hmw_analysis_sectors()
}
