% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{add_hmw_region_codes}
\alias{add_hmw_region_codes}
\title{Add the regional codes used on analysis of human muscle work.}
\usage{
add_hmw_region_codes(
  .df,
  concordance_path = MWTools::fao_concordance_path(),
  country_col = MWTools::conc_cols$country_col,
  country_code_col = MWTools::conc_cols$country_code_col,
  hmw_region_code_col = MWTools::conc_cols$hmw_region_code_col,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  sector_col = MWTools::mw_constants$sector_col,
  year = MWTools::mw_constants$year,
  yearly_working_hours_ilo_col = MWTools::ilo_cols$yearly_working_hours_ilo_col,
  employed_persons_ilo_col = MWTools::ilo_cols$employed_persons_ilo_col
)
}
\arguments{
\item{concordance_path}{The path to the country code concordance information.
Set to the bundled information by default,
retrieved using the \code{fao_concordance_path} function.}

\item{country_col, hmw_region_code_col}{See \code{MWTools::conc_cols}.}
}
\value{

}
\description{
...
}
