% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_row_col_meta}
\alias{add_row_col_meta}
\title{Prep PFU data for conversion to PSUT matrices}
\usage{
add_row_col_meta(
  .df,
  biomass = MWTools::mw_products$biomass,
  food = MWTools::mw_products$food,
  feed = MWTools::mw_products$feed,
  hu_mech = MWTools::mw_products$hu_mech,
  an_mech = MWTools::mw_products$an_mech,
  an_p = MWTools::mw_products$an_p,
  resources = MWTools::mw_sectors$resources_sector,
  farms = MWTools::mw_sectors$farms,
  food_production = MWTools::mw_sectors$food_production,
  feed_production = MWTools::mw_sectors$feed_production,
  human_females = MWTools::mw_species$human_females,
  human_males = MWTools::mw_species$human_males,
  transport = MWTools::mw_sectors$transport_sector,
  final = MWTools::all_stages$final,
  useful = MWTools::all_stages$useful,
  species = MWTools::mw_constants$species,
  sector = MWTools::mw_constants$sector_col,
  product = MWTools::mw_constants$product,
  last_stage = MWTools::mw_constants$last_stage,
  matnames = MWTools::mat_meta_cols$matnames,
  rownames = MWTools::mat_meta_cols$rownames,
  colnames = MWTools::mat_meta_cols$colnames,
  rowtypes = MWTools::mat_meta_cols$rowtypes,
  coltypes = MWTools::mat_meta_cols$coltypes,
  R_name = MWTools::psut_cols$R,
  U_name = MWTools::psut_cols$U,
  V_name = MWTools::psut_cols$V,
  Y_name = MWTools::psut_cols$Y,
  industry_type = MWTools::row_col_types$industry,
  product_type = MWTools::row_col_types$product,
  product_notation = RCLabels::from_notation,
  resource_notation = RCLabels::of_notation,
  species_notation = RCLabels::arrow_notation
)
}
\arguments{
\item{.df}{A data frame containing muscle work data,
likely the output from \code{specify_last_stages()}.}
}
\value{
A data frame containing PSUT matrices.
}
\description{
Primary-final-useful (PFU) data need to be converted to PSUT matrices.
This function take the output of \code{specify_last_stages()}
and converts to PSUT matrices in a data frame.
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu()
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu()
specify_product(hmw_df, amw_df) \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines() \%>\%
  specify_last_stages() \%>\%
  add_row_col_meta()
}
