% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{amw_analysis_data_path}
\alias{amw_analysis_data_path}
\title{Gives a file path to the animal muscle work analysis excel workbook}
\usage{
amw_analysis_data_path()
}
\value{
a path to to the animal muscle work analysis excel workbook bundled with this package
}
\description{
Gives a file path to the animal muscle work analysis excel workbook
}
\examples{
amw_analysis_data_path()
}
