% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{amw_data_path}
\alias{amw_data_path}
\title{Gives a file path to the FAO live animals data}
\usage{
amw_data_path()
}
\value{
a path to the bundled FAO data for the number of live animals
}
\description{
Gives a file path to the FAO live animals data
}
\examples{
amw_data_path()
}
