% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{amw_test_data_path}
\alias{amw_test_data_path}
\title{Gives a file path to the FAO animal muscle work test data}
\usage{
amw_test_data_path()
}
\value{
a path to to the FAO animal muscle work test data bundled with this package
}
\description{
Gives a file path to the FAO animal muscle work test data
}
\examples{
amw_test_data_path()
}
