% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{append_S_units_col}
\alias{append_S_units_col}
\title{Add an S_units column given other columns of PSUT matrices}
\usage{
append_S_units_col(
  .df,
  unit = MWTools::mw_cols$unit,
  product = MWTools::mw_cols$product,
  s_units = MWTools::psut_cols$s_units,
  product_notation = RCLabels::from_notation,
  unit_type = MWTools::row_col_types$unit
)
}
\arguments{
\item{.df}{A data frame}

\item{unit, product}{See \code{MWTools::mw_cols}.}

\item{s_units}{See \code{MWTools::psut_cols}.}

\item{product_notation}{Notation for products. Default is \code{RCLabels::from_notation}.}

\item{unit_type}{The type for units columns. Default is \code{MWTools::row_col_types$unit}.}
}
\value{
\code{.df} with an \code{s_units} column added.
}
\description{
The \code{S_units} column is a column of unit summation vectors
with products (energy carriers) in rows and units in columns.
The row names for each vector are the various products in that row of \code{.df}.
The column names for each vector are taken from the \code{units} column of \code{.df}.
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu() \%>\%
  # Keep only a few years for speed.
  dplyr::filter(Year \%in\% 2000:2002)
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu() \%>\%
  # Keep only a few years for speed.
  dplyr::filter(Year \%in\% 2000:2002)
specify_energy_type_method(hmw_df, amw_df) \%>\%
  specify_product() \%>\%
  specify_ktoe() \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines() \%>\%
  specify_last_stages() \%>\%
  MWTools::add_row_col_meta() \%>\%
  MWTools::collapse_to_psut() \%>\%
  append_S_units_col()
}
