% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{append_U_feed_U_eiou_r_eiou_cols}
\alias{append_U_feed_U_eiou_r_eiou_cols}
\title{Add U_feed, U_eiou, and r_eiou columns to a muscle work PSUT data frame}
\usage{
append_U_feed_U_eiou_r_eiou_cols(
  .df,
  U = MWTools::psut_cols$U,
  U_feed = MWTools::psut_cols$U_feed,
  U_eiou = MWTools::psut_cols$U_eiou,
  r_eiou = MWTools::psut_cols$r_eiou
)
}
\arguments{
\item{.df}{A PSUT data frame containing a column of \code{U} matrices.}

\item{U, U_feed, U_eiou, r_eiou}{See \code{MWTools::psut_cols}.}
}
\value{
\code{.df} with \code{U_feed}, \code{U_eiou}, and \code{r_eiou} matrices added.
}
\description{
\code{U_feed} is simply a copy of \code{U}.
\code{U_eiou} is a copy of \code{U} with \code{0} entries everywhere.
\code{r_eiou} has same structure as \code{U} but contains all \code{1}'s.
}
