% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_amw_numbers}
\alias{calc_amw_numbers}
\title{Calculate the number of live and working animals by country, species, and sector}
\usage{
calc_amw_numbers(.df)
}
\arguments{
\item{data_path}{The path to the folder containing data for the number of live
animals. Usually downloaded with the function \code{down_fao_live_animals}.}
}
\value{

}
\description{
Calculate the number of live and working animals by country, species, and sector
and over time. This function acts as a helper function calling a number of functions
in sequence to convert FAO data for live animals, usually downloaded with the function
\code{down_fao_live_animals}, into a tidy data frame.
}
\examples{
tidy_amw_numbers_data <- <- MWTools::down_fao_live_animals(data_path = file.path(fs::home_path(), "FAO_data")) \%>\%
  calc_amw_numbers(data_path = file.path(fs::home_path(), "FAO_data")

}
