% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{calc_hmw_primary_energy}
\alias{calc_hmw_primary_energy}
\title{Title}
\usage{
calc_hmw_primary_energy(
  .df,
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path(),
  year = MWTools::mw_constants$year,
  unit = MWTools::mw_constants$unit,
  exemplar_method_col = MWTools::mw_constants$exemplar_method_col,
  hmw_harvest_waste_sheet = MWTools::hmw_analysis_constants$hmw_harvest_waste_sheet,
  final_energy_col = MWTools::hmw_analysis_constants$final_energy_col,
  primary_energy_col = MWTools::hmw_analysis_constants$primary_energy_col,
  hmw_harvest_waste_col = MWTools::hmw_analysis_constants$hmw_harvest_waste_col,
  hmw_region_code_col = MWTools::conc_cols$hmw_region_code_col
)
}
\arguments{
\item{hmw_harvest_waste_col}{}
}
\value{

}
\description{
Title
}
