% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{calc_hmw_useful_energy}
\alias{calc_hmw_useful_energy}
\title{Title}
\usage{
calc_hmw_useful_energy(
  .df,
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path(),
  sector_col = MWTools::mw_constants$sector_col,
  year = MWTools::mw_constants$year,
  unit = MWTools::mw_constants$unit,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  hmw_power_sheet = MWTools::hmw_analysis_constants$hmw_power_sheet,
  hmw_analysis_sector_col = MWTools::hmw_analysis_constants$hmw_analysis_sector_col,
  power_col = MWTools::hmw_analysis_constants$power_col,
  hmw_region_code_col = MWTools::conc_cols$hmw_region_code_col,
 
    total_working_hours_ilo_col = MWTools::hmw_analysis_constants$total_working_hours_ilo_col,
  useful_energy_hmw_col = MWTools::hmw_analysis_constants$useful_energy_hmw_col,
  hours_to_seconds = MWTools::unit_constants$hours_to_seconds,
  joules_to_megajoules = MWTools::unit_constants$joules_to_megajoules
)
}
\arguments{
\item{.df}{}
}
\value{

}
\description{
Title
}
