% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_primary_energy}
\alias{calc_primary_energy}
\title{Title}
\usage{
calc_primary_energy(
  .df,
  harvest_waste = 0.45,
  final_energy_total = MWTools::amw_analysis_constants$final_energy_total,
  final_energy_ag = MWTools::amw_analysis_constants$final_energy_ag,
  final_energy_tr = MWTools::amw_analysis_constants$final_energy_tr,
  primary_energy_total = MWTools::amw_analysis_constants$primary_energy_total,
  primary_energy_ag = MWTools::amw_analysis_constants$primary_energy_ag,
  primary_energy_tr = MWTools::amw_analysis_constants$primary_energy_tr
)
}
\arguments{
\item{primary_energy_tr}{}
}
\value{

}
\description{
Title
}
