% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{calc_total_hours_worked}
\alias{calc_total_hours_worked}
\title{Calculate the total number of hours worked each year}
\usage{
calc_total_hours_worked(
  .df,
  yearly_working_hours_ilo_col = MWTools::ilo_cols$yearly_working_hours_ilo_col,
  employed_persons_ilo_col = MWTools::ilo_cols$employed_persons_ilo_col,
 
    total_working_hours_ilo_col = MWTools::hmw_analysis_constants$total_working_hours_ilo_col
)
}
\arguments{
\item{yearly_working_hours_ilo_col, employed_persons_ilo_col}{See \code{MWTools::ilo_cols}.}

\item{total_working_hours_ilo_col}{See \code{MWTools::hmw_analysis_constants}}
}
\value{

}
\description{
Calc...
}
