% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_useful_energy}
\alias{calc_useful_energy}
\title{Title}
\usage{
calc_useful_energy(
  .df,
  amw_analysis_path,
  species = MWTools::mw_constants$species,
  wa_power_sheet = MWTools::amw_analysis_constants$wa_power_sheet,
  wa_days_hours_sheet = MWTools::amw_analysis_constants$wa_days_hours_sheet,
  working_days_col = MWTools::amw_analysis_constants$working_days_col,
  working_hours_col = MWTools::amw_analysis_constants$working_hours_col,
  working_seconds_col = MWTools::amw_analysis_constants$working_seconds_col,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  method_source = MWTools::amw_analysis_constants$method_source,
  power_per_animal = MWTools::amw_analysis_constants$power_per_animal,
  useful_energy_total = MWTools::amw_analysis_constants$useful_energy_total,
  useful_energy_ag = MWTools::amw_analysis_constants$useful_energy_ag,
  useful_energy_tr = MWTools::amw_analysis_constants$useful_energy_tr,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  working_animals_ag_col = MWTools::amw_analysis_constants$working_animals_ag_col,
  working_animals_tr_col = MWTools::amw_analysis_constants$working_animals_tr_col
)
}
\arguments{
\item{working_animals_tr_col}{}
}
\value{

}
\description{
Title
}
