% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_working_animals}
\alias{calc_working_animals}
\title{Calculate the number of working animals}
\usage{
calc_working_animals(
  .df,
  amw_analysis_path = MWTools::amw_analysis_data_path(),
  year = MWTools::mw_constants$year,
  species = MWTools::mw_constants$species,
  prop_working_animals_col = MWTools::amw_analysis_constants$prop_working_animals_col,
  wa_perc_sheet = MWTools::amw_analysis_constants$wa_perc_sheet,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  live_animals_col = MWTools::amw_analysis_constants$live_animals_col,
  amw_region_code_col = MWTools::conc_cols$amw_region_code_col,
  amw_region_col = MWTools::amw_analysis_constants$amw_region_col,
  exemplar_method_col = MWTools::amw_analysis_constants$exemplar_method_col
)
}
\arguments{
\item{.df}{A data frame ...}

\item{amw_analysis_path}{The path to the animal muscle work analysis data,
containing data for the proportion of working animals
by species, country, and over time. Set to the function
\code{MWTools::amw_analysis_data_path()}, by default,
which returns the path the analysis data bundled
with the \code{MWTools} package.}

\item{prop_working_animal, da_perc, working_animals_total_col, live_animals_col, mw_region_code_col, mw_region_col, exemplar_method_col}{See \code{MWTools::amw_analysis_constants}.}

\item{...}{}
}
\value{

}
\description{
Calculate the number of working animals using data for the number of live
animals from the FAO, and user-supplied data for the proportion of working
animals by region and year.
}
\examples{
working_animals_data <- MWTools::down_fao_live_animals(data_path = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_path = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes() \%>\%
  trim_fao_data() \%>\%
  get_working_species() \%>\%
  calc_working_animals()

}
