% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_yearly_feed}
\alias{calc_yearly_feed}
\title{Title}
\usage{
calc_yearly_feed(
  .df,
  amw_analysis_path = MWTools::amw_analysis_data_path(),
  species = MWTools::mw_constants$species,
  wa_feed_sheet = MWTools::amw_analysis_constants$wa_feed_sheet,
  wa_days_hours_sheet = MWTools::amw_analysis_constants$wa_days_hours_sheet,
  working_days_col = MWTools::amw_analysis_constants$working_days_col,
  nonworking_days_col = MWTools::amw_analysis_constants$nonworking_days_col,
  working_hours_col = MWTools::amw_analysis_constants$working_hours_col,
  working_day_feed_col = MWTools::amw_analysis_constants$working_day_feed_col,
  nonworking_day_feed_col = MWTools::amw_analysis_constants$nonworking_day_feed_col,
  working_yearly_feed_col = MWTools::amw_analysis_constants$working_yearly_feed_col,
 
    nonworking_yearly_feed_col = MWTools::amw_analysis_constants$nonworking_yearly_feed_col,
  total_yearly_feed_col = MWTools::amw_analysis_constants$total_yearly_feed_col,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  method_source = MWTools::amw_analysis_constants$method_source
)
}
\arguments{
\item{method_source}{}
}
\value{

}
\description{
Title
}
