% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conc_cols}
\alias{conc_cols}
\title{Country concordance columns}
\format{
A string list with 8 entries.
\describe{
\item{country_code_col}{The metadata column name "Country.code" containing three letter ISO 3166-1 country codes.}
\item{country_col}{The metadata column name "Country" containing three letter ISO 3166-1 country codes.}
\item{country_code_pfu_col}{The metadata column name "Country_PFU" containing bespoke three letter country codes.}
\item{country_incl_col}{The metadata column name "Country.incl." containing either "Yes" or "No" depending on whether a particulr region or country shoule be included for analysis in \code{MWTools}.}
\item{amw_region_code_col}{The metadata column "AMW.Region.code" telling the region for this row.}
\item{hmw_region_code_col}{The metadata column "HMW.Region.code" telling the region for this row.}
\item{mapping_sheet}{The name of the sheet for name mappings.}
}
}
\usage{
conc_cols
}
\description{
A string list containing the names of the columns containing country code information.
}
\examples{
conc_cols
}
\keyword{datasets}
