% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_amw_data_functions.R
\name{down_fao_live_animals}
\alias{down_fao_live_animals}
\title{Download live animals data from FAOSTAT}
\usage{
down_fao_live_animals(
  data_folder = data_folder,
  live_animals_code = MWTools::fao_codes$live_animals_code
)
}
\arguments{
\item{data_folder}{A user-specified file path to the directory in which to
download the FAO live animals data.}

\item{live_animals_code}{See \code{MWTools::fao_codes$live_animals_code}.}
}
\value{

}
\description{
This function uses the bulk download facility function from the R package
\code{FAOSTAT} to download data for the number of live animals by country over time.
}
\examples{
live_animals_data <- down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data"))

}
