% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_types}
\alias{energy_types}
\title{Energy types in muscle work data frames}
\format{
A string list with 2 entries.
\describe{
See \code{IEATools::energy_types}.
}
}
\usage{
energy_types
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
energy_types
}
\keyword{datasets}
