% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extdata_path}
\alias{extdata_path}
\title{Gives a file path to the extdata directory}
\usage{
extdata_path()
}
\value{
a path to to the concordance data excel workbook bundled with this package
}
\description{
Gives a file path to the extdata directory
}
\examples{
fao_concordance_path()
}
