% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_codes}
\alias{fao_codes}
\title{FAO data item code names}
\format{
A string list with 1 entries.
\describe{
\item{live_animals_code}{The string "QA" representing data for the number of live animals.}
}
}
\usage{
fao_codes
}
\description{
A string list containing a selection of FAO data code names. See \code{FAOSTAT::FAOsearch()}
to view the associated items, codes, and other metadata.
}
\examples{
fao_codes
}
\keyword{datasets}
