% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_cols}
\alias{fao_cols}
\title{FAO data column names}
\format{
A string list with 5 entries.
\describe{
\item{area_fao_col}{The string "area".}
\item{item_fao_col}{The string "item".}
\item{year_fao_col}{The string "year".}
\item{unit_fao_col}{The string "unit".}
\item{value_fao_col}{The string "value".}
}
}
\usage{
fao_cols
}
\description{
A string list containing a selection of FAO data column names. For data retrieved by the package \code{FAOSTAT}.
}
\examples{
fao_cols
}
\keyword{datasets}
