% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fao_concordance_path}
\alias{fao_concordance_path}
\title{Gives a file path to the concordance data excel workbook}
\usage{
fao_concordance_path()
}
\value{
a path to to the concordance data excel workbook bundled with this package
}
\description{
Gives a file path to the concordance data excel workbook
}
\examples{
fao_concordance_path()
}
