% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{fill_ilo_data}
\alias{fill_ilo_data}
\title{Fill missing data for the number of hours worked and employed persons based on the earliest value}
\usage{
fill_ilo_data(
  .df,
  country_col = MWTools::conc_cols$country_col,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  sector_col = MWTools::mw_constants$sector_col,
  year = MWTools::mw_cols$year,
  yearly_working_hours_ilo_col = MWTools::ilo_cols$yearly_working_hours_ilo_col,
  employed_persons_ilo_col = MWTools::ilo_cols$employed_persons_ilo_col
)
}
\arguments{
\item{.df}{The ILO labor data with added region codes.
Usually produced by calling the
\code{add_hmw_region_codes} function in sequence on the raw FAO data.}

\item{country_col}{See \code{MWTools::conc_cols}.}

\item{sex_ilo_col, yearly_working_hours_ilo_col, employed_persons_ilo_col}{See \code{MWTools::ilo_cols}.}

\item{sector_col, year}{See \code{MWTools::mw_constants}.}
}
\description{
Fill missing data for the number of hours worked and employed persons based on the earliest value
}
\examples{
working_humans_data <- read.csv(file = MWTools::hmw_test_data_path()) \%>\%
  add_hmw_region_codes() \%>\%
  fill_ilo_data()
}
