% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{get_aggregate_data}
\alias{get_aggregate_data}
\title{Retrieve data with sectors organised by "Aggregate"}
\usage{
get_aggregate_data(.df)
}
\arguments{
\item{.df}{}
}
\value{

}
\description{
Retrieve data with sectors organised by "Aggregate"
}
\examples{

}
