% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hmw_analysis_data_path}
\alias{hmw_analysis_data_path}
\title{Gives a file path to the human muscle work analysis excel workbook}
\usage{
hmw_analysis_data_path()
}
\value{
a path to to the human muscle work analysis excel workbook bundled with this package
}
\description{
Gives a file path to the human muscle work analysis excel workbook
}
\examples{
hmw_analysis_data_path()
}
