% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmw_sector_constants}
\alias{hmw_sector_constants}
\title{ILO sector name constants}
\format{
A string list with 6 entries.
\describe{
\item{ilo_agr_name}{The string name for the agriculture sector in the ILO data.}
\item{ilo_ind_name}{The string name for the industry sector in the ILO data.}
\item{ilo_ser_name}{The string name for the services sector in the ILO data.}
\item{agr_name}{The string name of the agriculture sector in the CL-PFU database, namely "Agriculture".}
\item{ind_name}{The string name of the industrial sector in the CL-PFU database, namely "Industry".}
\item{ser_name}{The string name of the services sector in the CL-PFU database, namely "Services".}
}
}
\usage{
hmw_sector_constants
}
\description{
A string list containing the names of relevant ILO sectors.
}
\examples{
hmw_sector_constants
}
\keyword{datasets}
