% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmw_sex_constants}
\alias{hmw_sex_constants}
\title{ILO sex name constants}
\format{
A string list with 8 entries.
\describe{
\item{ilo_female_name}{The string name identifying the female sex in the ILO data.}
\item{ilo_male_name}{The string name identifying the male sex in the ILO data.}
\item{ilo_other_name}{The string name identifying the other sex in the ILO data.}
\item{ilo_total_name}{The string name identifying the total sex in the ILO data.}
\item{female_name}{The string name identifying the female sex in the CL-PFU database, namely "Female".}
\item{male_name}{The string name identifying the male sex in the CL-PFU database, namely "Male".}
\item{other_name}{The string name identifying the other sex in the CL-PFU database, namely "Other".}
\item{total_name}{The string name identifying the total sex in the CL-PFU database, namely "Total".}
}
}
\usage{
hmw_sex_constants
}
\description{
A string list containing the names of relevant ILO sex identifiers.
}
\examples{
hmw_sex_constants
}
\keyword{datasets}
