% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hmw_test_data_path}
\alias{hmw_test_data_path}
\title{Gives a file path to the ILO human muscle work test data}
\usage{
hmw_test_data_path()
}
\value{
a path to to the ILO human muscle work test data bundled with this package
}
\description{
Gives a file path to the ILO human muscle work test data
}
\examples{
hmw_test_data_path()
}
