% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ilo_codes}
\alias{ilo_codes}
\title{ILO data item code names}
\format{
A string list with 2 entries.
\describe{
\item{working_hours_code}{The string "HOW_TEMP_SEX_ECO_NB_A", representing Mean weekly hours actually worked per employed person by sex and economic activity.}
\item{employment_code}{The string "EMP_TEMP_SEX_ECO_NB_A", representing Employment by sex and economic activity (thousands).}
}
}
\usage{
ilo_codes
}
\description{
A string list containing a selection of ILO data code names. See \code{Rilostat::get_ilostat_toc()}
to view the associated items, codes, and other metadata..
}
\examples{
ilo_codes
}
\keyword{datasets}
