% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ilo_workingHours_test_data_path}
\alias{ilo_workingHours_test_data_path}
\title{Gives a file path to the ILO working hours test data}
\usage{
ilo_workingHours_test_data_path()
}
\value{
a path to to the ILO working hours test data bundled with this package
}
\description{
Gives a file path to the ILO working hours test data
}
\examples{
ilo_workingHours_test_data_path()
}
