% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_stages}
\alias{last_stages}
\title{Last stages for energy conversion chains in PSUT format}
\format{
A string list with 3 entries.
\describe{
See \code{IEATools::last_stages}.
}
}
\usage{
last_stages
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
last_stages
}
\keyword{datasets}
