% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_meta_cols}
\alias{mat_meta_cols}
\title{Names of matrix metadata columns in muscle work data frames}
\format{
A string list with 6 entries.
\describe{
See \code{IEATools::mat_meta_cols}.
}
}
\usage{
mat_meta_cols
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
mat_meta_cols
}
\keyword{datasets}
