% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Methods of attributing primary energy from final energy in muscle work data frames}
\format{
A string list with 3 entries.
\describe{
See \code{IEATools::methods}.
}
}
\usage{
methods
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
methods
}
\keyword{datasets}
