% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_cols}
\alias{mw_cols}
\title{Column names in muscle work data frames}
\format{
A string list with 11 entries.
\describe{
See \code{IEATools::iea_cols}.
}
}
\usage{
mw_cols
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
mw_cols
}
\keyword{datasets}
