% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_products}
\alias{mw_products}
\title{Muscle work products}
\format{
A string list with 6 entries.
\describe{
\item{food}{The name for food final energy.}
\item{feed}{The name for feed final energy.}
\item{biomass}{The name for biomass primary energy.}
\item{hu_mech}{The name for human mechanical work useful energy.}
\item{an_mech}{The name for animal mechanical work useful energy.}
\item{an_p}{The name for animal propulsion work useful energy.}
}
}
\usage{
mw_products
}
\description{
A list containing names of muscle work products (energy carriers).
}
\examples{
mw_products
}
\keyword{datasets}
