% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_species}
\alias{mw_species}
\title{Working animal species}
\format{
A string list with 9 entries.
\describe{
\item{asses}{The string "Asses".}
\item{camels}{The string "Camels".}
\item{cattle}{The string "Cattle".}
\item{horses}{The string "Horses".}
\item{mules}{The string "Mules".}
\item{buffaloes}{The string "Buffaloes".}
\item{camelids_other}{The string "Camelids, other".}
\item{camelids}{The string "Camelids", representing the combined category of "Camels" and "Camelids, other".}
}
}
\usage{
mw_species
}
\description{
A string list containing the species of animals considered to perform muscle work
}
\examples{
mw_species
}
\keyword{datasets}
