% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_constants_mw}
\alias{phi_constants_mw}
\title{Muscle work phi constants}
\format{
A data frame with 6 rows. Each row provides the phi value for one energy carrier:
\describe{
\item{Product}{A string column of muscle work energy carriers.}
\item{phi}{A string column of exergy-to-energy ratios.}
\item{is.useful}{A boolean column telling whether this energy carrier is found at the useful stage (or not).}
}
}
\usage{
phi_constants_mw
}
\description{
A data frame containing
- muscle work products (energy carriers),
- phi (exergy-to-energy ratio) values,
- and an "is.useful" column saying whether that energy carrier is found
at the useful stage.
}
\examples{
phi_constants_mw
}
\keyword{datasets}
