% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_psut}
\alias{prep_psut}
\title{Convert a tidy data frame to PSUT matrices}
\usage{
prep_psut(
  .df,
  country = MWTools::mw_cols$country,
  year = MWTools::mw_cols$year,
  method = MWTools::mw_cols$method,
  energy_type = MWTools::mw_cols$energy_type,
  last_stage = MWTools::mw_cols$last_stage,
  unit = MWTools::mw_cols$unit,
  e_dot = MWTools::mw_cols$e_dot,
  matnames = MWTools::mat_meta_cols$matnames,
  matvals = MWTools::mat_meta_cols$matvals,
  rownames = MWTools::mat_meta_cols$rownames,
  colnames = MWTools::mat_meta_cols$colnames,
  rowtypes = MWTools::mat_meta_cols$rowtypes,
  coltypes = MWTools::mat_meta_cols$coltypes
)
}
\arguments{
\item{country, year, method, energy_type, last_stage, unit, e_dot}{See \code{MWTools::mw_cols}.}

\item{matnames, matvals, rownames, colnames, rowtypes, coltypes}{See \code{MWTools::mat_meta_cols}.}
}
\value{
A \code{matsindf}-style, wide-by-matrices data frame of muscle work matrices.
}
\description{
A tidy data frame of muscle work information can be converted to
a \code{matsindf} data frame via this function.
}
\details{
Prior to forming matrices, this function deletes unneeded columns
(columns that are neither metadata nor energy values).
It also aggregates data frame rows that will end up at the same
row, column location in the matrices.
}
