% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psut_cols}
\alias{psut_cols}
\title{Column names in PSUT data frames}
\format{
A string list with 16 entries.
\describe{
See \code{IEATools::psut_cols}.
}
}
\usage{
psut_cols
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
psut_cols
}
\keyword{datasets}
