% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{row_col_types}
\alias{row_col_types}
\title{Row and column types for matrices made from muscle work data frames}
\format{
A string list with 6 entries.
\describe{
See \code{IEATools::row_col_types}.
}
}
\usage{
row_col_types
}
\description{
A string list containing package constants used in MWTools package functions.
This list is borrowed directly from the \code{IEATools} package.
}
\examples{
row_col_types
}
\keyword{datasets}
