% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_fu_machines}
\alias{specify_fu_machines}
\title{Specify final-to-useful machines (species)}
\usage{
specify_fu_machines(
  .df,
  product = MWTools::mw_constants$product,
  stage = MWTools::mw_constants$stage_col,
  species = MWTools::mw_constants$species,
  useful = MWTools::all_stages$useful,
  product_notation = RCLabels::from_notation,
  machine_notation = RCLabels::arrow_notation
)
}
\arguments{
\item{.df}{A data frame, likely the output of \code{specify_useful_products()}.}
}
\value{

}
\description{
Final-to-useful machines should be named as "Machine -> Product".
This function renames final-to-useful machines (species) to the desired form.
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu()
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu()
specify_product(hmw_df, amw_df) \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines()
}
