% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_ktoe}
\alias{specify_ktoe}
\title{Converts units from EJ to ktoe}
\usage{
specify_ktoe(
  .df,
  energy_col = MWTools::mw_constants$energy_col,
  units_col = MWTools::mw_constants$units_col
)
}
\arguments{
\item{.df}{A data frame with \code{units} and \code{energy} columns.}
}
\value{
\code{.df} with \code{energy} column converted from EJ to ktoe.
}
\description{
Primary-final-useful (PFU) data from \code{calc_hmw_pfu()} and \code{calc_amw_pfu()}
are in EJ.
This function converts to ktoe.
}
\examples{
hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu() \%>\%
  specify_ktoe()
}
