% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_last_stages}
\alias{specify_last_stages}
\title{Create energy conversion chains for final and useful last stages}
\usage{
specify_last_stages(
  .df,
  last_stage = MWTools::mw_constants$last_stage,
  final = MWTools::last_stages$final,
  useful = MWTools::last_stages$useful
)
}
\arguments{
\item{.df}{A data frame, probably the output of \code{specify_fu_machines()}.}
}
\value{
A data frame containing a new column \code{last_stage} and
rows for both final and useful being the last stage.
}
\description{
Create energy conversion chains for final and useful last stages
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu()
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu()
specify_product(hmw_df, amw_df) \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines() \%>\%
  specify_last_stages()
}
