% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{tidy_hmw_pfu}
\alias{tidy_hmw_pfu}
\title{Title}
\usage{
tidy_hmw_pfu(
  .df,
  year = MWTools::mw_constants$year,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  sector_col = MWTools::mw_constants$sector_col,
  species = MWTools::mw_constants$species,
  country_col = MWTools::conc_cols$country_col,
  hmw_region_code_col = MWTools::conc_cols$hmw_region_code_col,
  final_energy_col = MWTools::hmw_analysis_constants$final_energy_col,
  primary_energy_col = MWTools::hmw_analysis_constants$primary_energy_col,
  useful_energy_hmw_col = MWTools::hmw_analysis_constants$useful_energy_hmw_col,
  hmw_analysis_sector_col = MWTools::hmw_analysis_constants$hmw_analysis_sector_col,
  energy_col = MWTools::mw_constants$energy_col,
  stage_col = MWTools::mw_constants$stage_col,
  units_col = MWTools::mw_constants$units_col
)
}
\arguments{
\item{stage_col}{}
}
\value{

}
\description{
Title
}
