% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{tidy_ilo_data}
\alias{tidy_ilo_data}
\title{Get tidy employment and working hours data for agriculture, industry, and services}
\usage{
tidy_ilo_data(.df)
}
\arguments{
\item{.df}{}
}
\value{

}
\description{
Get tidy employment and working hours data for agriculture, industry, and services
}
\examples{

}
