% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{tidy_mw_df}
\alias{tidy_mw_df}
\title{Prep PFU data for conversion to PSUT matrices}
\usage{
tidy_mw_df(.pfu_df)
}
\arguments{
\item{.pfu_df}{A PFU data frame,
the output from \code{specify_primary_production()}.}
}
\value{
A data frame with correct rows (many duplicated) and a matrix name column.
}
\description{
Primary-final-useful (PFU) data need to be prepared
for conversion to PSUT matrices by
duplicating some rows and adding a matrix name column.
This function performs that preparation.
}
