% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{tidy_numbers_data}
\alias{tidy_numbers_data}
\title{Title}
\usage{
tidy_numbers_data(
  .df,
  amw_analysis_path,
  year = MWTools::mw_constants$year,
  species = MWTools::mw_constants$species,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  country_code_col = MWTools::conc_cols$country_code_col,
  working_animals_col = MWTools::amw_analysis_constants$working_animals_col,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  working_animals_ag_col = MWTools::amw_analysis_constants$working_animals_ag_col,
  working_animals_tr_col = MWTools::amw_analysis_constants$working_animals_tr_col,
  live_animals_col = MWTools::amw_analysis_constants$live_animals_col,
  sector_col = MWTools::mw_constants$sector_col
)
}
\arguments{
\item{sector_col}{}
}
\value{

}
\description{
Applied to a data frame after calling the functions...
}
