% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{tidy_pfu_data}
\alias{tidy_pfu_data}
\title{Title}
\usage{
tidy_pfu_data(
  .df,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  country_code_col = MWTools::conc_cols$country_code_col,
  year = MWTools::mw_constants$year,
  species = MWTools::mw_constants$species,
  useful_energy_total = MWTools::amw_analysis_constants$useful_energy_total,
  useful_energy_ag = MWTools::amw_analysis_constants$useful_energy_ag,
  useful_energy_tr = MWTools::amw_analysis_constants$useful_energy_tr,
  final_energy_total = MWTools::amw_analysis_constants$final_energy_total,
  final_energy_ag = MWTools::amw_analysis_constants$final_energy_ag,
  final_energy_tr = MWTools::amw_analysis_constants$final_energy_tr,
  primary_energy_total = MWTools::amw_analysis_constants$primary_energy_total,
  primary_energy_ag = MWTools::amw_analysis_constants$primary_energy_ag,
  primary_energy_tr = MWTools::amw_analysis_constants$primary_energy_tr,
  sector_col = MWTools::mw_constants$sector_col,
  stage_col = MWTools::mw_constants$stage_col,
  energy_mj_year = MWTools::mw_constants$energy_mj_year
)
}
\arguments{
\item{primary_energy_tr}{}
}
\value{

}
\description{
Title
}
