% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_amw_data_functions.R
\name{add_concordance_codes}
\alias{add_concordance_codes}
\title{Add 3 letter ISO codes and MWTools-specific region codes}
\usage{
add_concordance_codes(
  .df,
  concordance_path = MWTools::fao_concordance_path(),
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  country_name = MWTools::mw_constants$country_name,
  country_incl_col = MWTools::conc_cols$country_incl_col,
  country_code_col = MWTools::conc_cols$country_code_col,
  country_code_pfu_col = MWTools::conc_cols$country_code_pfu_col
)
}
\arguments{
\item{.df}{The data frame containing tidy live animals data, to which to add
3 letter ISO codes.}

\item{concordance_path}{The file path to concordance information mapping the
FAO country names supplied in FAOSTAT to 3-letter ISO
codes, and MWTools specific region codes. Set to
\code{MWTools::fao_concordance_path()} by default, the path
to the bundled concordance information in \code{MWTools}.}

\item{mw_region_code_col}{See \code{MWTools::amw_analysis_constants}.}

\item{country_name}{See \code{MWTools::mw_constants}.}

\item{country_incl_col, country_code_col, country_code_pfu_col}{See \code{MWTools::conc_cols}.}
}
\value{

}
\description{
Adds 3 letter ISO country codes and muscle work region codes for countries
present in the FAO's Live animals data, downloaded through the \code{FAOSTAT} package,
usually using the \code{MWTools::down_fao_live_animals} function.
}
\examples{
live_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes()

}
