% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{add_hmw_analysis_sectors}
\alias{add_hmw_analysis_sectors}
\title{Add the human muscle work sectors used in the bundled analysis data}
\usage{
add_hmw_analysis_sectors(
  .df,
  sector_col = MWTools::mw_constants$sector_col,
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path(),
  hmw_sector_map_sheet = MWTools::hmw_analysis_constants$hmw_sector_map_sheet,
  hmw_analysis_sector_col = MWTools::hmw_analysis_constants$hmw_analysis_sector_col,
  agriculture = MWTools::hmw_analysis_constants$agriculture_broad.sector,
  industry = MWTools::hmw_analysis_constants$industry_broad.sector,
  services = MWTools::hmw_analysis_constants$services_broad.sector
)
}
\arguments{
\item{sector_col}{See \code{MWTools::mw_constants}.}

\item{hmw_analysis_data_path}{See \code{MWTools::hmw_analysis_data_path()}.}

\item{hmw_sector_map_sheet, hmw_analysis_sector_col, agriculture, industry, services}{See \code{MWTools::hmw_analysis_constants}.}
}
\value{

}
\description{
...
}
\examples{

}
