% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{append_U_feed_U_eiou_r_eiou_cols}
\alias{append_U_feed_U_eiou_r_eiou_cols}
\title{Add U_feed, U_eiou, and r_eiou columns to a muscle work PSUT data frame}
\usage{
append_U_feed_U_eiou_r_eiou_cols(
  .df = NULL,
  U = MWTools::psut_cols$U,
  U_feed = MWTools::psut_cols$U_feed,
  U_eiou = MWTools::psut_cols$U_eiou,
  r_eiou = MWTools::psut_cols$r_eiou
)
}
\arguments{
\item{.df}{A PSUT data frame containing a column of \code{U} matrices.
Default is \code{NULL}, allowing a single matrix for the \code{U} argument.}

\item{U}{The name of the incoming \code{U} matrix. See \code{MWTools::psut_cols}.}

\item{U_feed, U_eiou, r_eiou}{Names for outgoing matrices. See \code{MWTools::psut_cols}.}
}
\value{
\code{.df} with new columns for \code{U_feed}, \code{U_eiou}, and \code{r_eiou} matrices.
}
\description{
\code{U_feed}, \code{U_eiou}, and \code{r_eiou} matrices are calculated from \code{U}.
All three matrices (\code{U_feed}, \code{U_eiou}, and \code{r_eiou})
have the same structure (row names, column names, row types, and column types)
as \code{U}.
For \code{MWTools}, there is no energy industry own use (EIOU),
so \code{U_feed} is simply a copy of \code{U}, and \code{U_eiou} and \code{r_eiou} are full of \code{0}s.
}
\details{
This function employs \code{matsindf::matsindf_apply()} internally, so
\code{U} can be either a single matrix or the name of the \code{U} column in \code{.df}.
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu() \%>\%
  # Keep only a few years for speed.
  dplyr::filter(Year \%in\% 2000:2002)
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu() \%>\%
  # Keep only a few years for speed.
  dplyr::filter(Year \%in\% 2000:2002)
with_U_cols <- specify_energy_type_method(hmw_df, amw_df) \%>\%
  specify_product() \%>\%
  specify_ktoe() \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines() \%>\%
  specify_last_stages() \%>\%
  MWTools::add_row_col_meta() \%>\%
  MWTools::collapse_to_psut() \%>\%
  append_S_units_col() \%>\%
  append_U_feed_U_eiou_r_eiou_cols()
}
