% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_amw_pfu}
\alias{calc_amw_pfu}
\title{Calculate primary, final, and useful working animal energy}
\usage{
calc_amw_pfu(.df)
}
\arguments{
\item{data_folder}{The file path to the FAO live animals data downloaded using
the function \code{MWTools::down_fao_live_animals}.}
}
\value{

}
\description{
Calculate the total number of working animals and primary, final, and useful
working animal energy by country, for six species: Asses, Buffaloes, Camelids,
Cattle, Horses, and Mules, and three sector categories: Total, Agriculture,
and Transport. This function acts as a helper function calling a number of functions
in sequence to convert FAO data for live animals, usually downloaded with the function
\code{down_fao_live_animals}, into a tidy data frame.
}
\examples{
tidy_amw_pfu_data <- <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  calc_amw_pfu(data_folder = file.path(fs::home_path(), "FAO_data")

}
