% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_amw_pfu}
\alias{calc_amw_pfu}
\title{Calculate primary, final, and useful working animal energy}
\usage{
calc_amw_pfu(amw_data_path, mw_mapping_path, amw_path)
}
\arguments{
\item{amw_path}{}
}
\value{

}
\description{
This function calculates the total number of working animals and primary,
final, and useful working animal energy by country, and for six species:
Asses, Buffaloes, Camelids, Cattle, Horses, and Mules.
}
\examples{
data <- calc_amw_pfu(amw_data_path, mw_mapping_path, amw_path) # Add paths here.

}
