% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_final_energy}
\alias{calc_final_energy}
\title{Title}
\usage{
calc_final_energy(
  .df,
  trough_waste = 0.1,
  ge_de_ratio = 1.16363636363636,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  working_animals_ag_col = MWTools::amw_analysis_constants$working_animals_ag_col,
  working_animals_tr_col = MWTools::amw_analysis_constants$working_animals_tr_col,
  total_yearly_feed_col = MWTools::amw_analysis_constants$total_yearly_feed_col,
  final_energy_total = MWTools::amw_analysis_constants$final_energy_total,
  final_energy_ag = MWTools::amw_analysis_constants$final_energy_ag,
  final_energy_tr = MWTools::amw_analysis_constants$final_energy_tr
)
}
\arguments{
\item{final_energy_tr}{}
}
\value{

}
\description{
Title
}
