% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{calc_hmw_final_energy}
\alias{calc_hmw_final_energy}
\title{Calculate the final energy consumed by human workers}
\usage{
calc_hmw_final_energy(
  .df,
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path(),
  sector_col = MWTools::mw_constants$sector_col,
  year = MWTools::mw_constants$year,
  unit = MWTools::mw_constants$unit,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  hmw_analysis_sector_col = MWTools::hmw_analysis_constants$hmw_analysis_sector_col,
  hmw_food_sheet = MWTools::hmw_analysis_constants$hmw_food_sheet,
  food_consumption_col = MWTools::hmw_analysis_constants$food_consumption_col,
 
    yearly_energy_consumption_pp_col = MWTools::hmw_analysis_constants$yearly_energy_consumption_pp_col,
  final_energy_col = MWTools::hmw_analysis_constants$final_energy_col,
  hmw_region_code_col = MWTools::conc_cols$hmw_region_code_col,
  kcal_to_mj = MWTools::unit_constants$kcal_to_mj,
  employed_persons_ilo_col = MWTools::ilo_cols$employed_persons_ilo_col
)
}
\arguments{
\item{hmw_analysis_data_path}{See \code{MWTools::hmw_analysis_data_path()}.}

\item{sector_col, year, unit}{See \code{MWTools::mw_constants}.}

\item{sex_ilo_col}{See \code{MWTools::ilo_cols}}

\item{hmw_analysis_sector_col, hmw_food_sheet, food_consumption_col}{See \code{MWTools::hmw_analysis_constants}.}

\item{hmw_region_code_col}{See \code{MWTools::conc_cols}.}
}
\value{

}
\description{
...
}
