% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{calc_hmw_pfu}
\alias{calc_hmw_pfu}
\title{Title}
\usage{
calc_hmw_pfu(
  .df,
  concordance_path = MWTools::fao_concordance_path(),
  hmw_analysis_data_path = MWTools::hmw_analysis_data_path()
)
}
\arguments{
\item{stage_col}{}
}
\value{

}
\description{
Title
}
