% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{calc_sector_split}
\alias{calc_sector_split}
\title{Calculate the number of working animals in agriculture and transport}
\usage{
calc_sector_split(
  .df,
  amw_analysis_path = MWTools::amw_analysis_data_path(),
  year = MWTools::mw_constants$year,
  species = MWTools::mw_constants$species,
  method_source = MWTools::amw_analysis_constants$method_source,
  metric = MWTools::amw_analysis_constants$metric,
  mw_region_col = MWTools::amw_analysis_constants$mw_region_col,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  wa_enduse_sheet = MWTools::amw_analysis_constants$wa_enduse_sheet,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  working_animals_ag_col = MWTools::amw_analysis_constants$working_animals_ag_col,
  working_animals_tr_col = MWTools::amw_analysis_constants$working_animals_tr_col,
 
    prop_working_animals_ag_col = MWTools::amw_analysis_constants$prop_working_animals_ag_col,
 
    prop_working_animals_tr_col = MWTools::amw_analysis_constants$prop_working_animals_tr_col
)
}
\arguments{
\item{.df}{A data frame containing live animals data...}

\item{amw_analysis_path}{The path to the animal muscle work analysis data,
containing data for the proportion of working animals
in agriculture and transport, by species, country, and
over time. Set to the function \code{MWTools::amw_analysis_data_path()},
by default, which returns the path the analysis data
bundled with the \code{MWTools} package.}

\item{year, species}{See \code{MWTools::mw_constants}.}

\item{method_source, metric, mw_region_col, mw_region_code_col, wa_enduse_sheet, working_animals_total_col, working_animals_ag_col, working_animals_tr_col, prop_working_animals_ag_col, prop_working_animals_tr_col}{See \code{MWTools::amw_analysis_constants}.}
}
\value{

}
\description{
Calculate the split of working animals performing muscle work in agriculture
and transport using user-supplied data for the proportion of animals working
in agriculture by region and year.
}
\examples{
working_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes() \%>\%
  trim_fao_data() \%>\%
  get_working_species() \%>\%
  calc_working_animals() \%>\%
  calc_sector_split()

}
