% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conc_cols}
\alias{conc_cols}
\title{Country concordance columns}
\format{
A string list with 9 entries.
\describe{
\item{country_code_col}{The metadata column name "Country.code" containing three letter ISO 3166-1 country codes.}
\item{country_col}{The metadata column name "Country" containing three letter ISO 3166-1 country codes.}
\item{country_incl_col}{The metadata column name "Country.incl." containing either "Yes" or "No" depending on whether a particular region or country should be included for analysis in \code{MWTools}.}
\item{exemplar_country_col}{The metadata column name "Exemplar.country", representing the country from which some data should be taken if absent from a particular country.}
\item{amw_region_code_col}{The metadata column "AMW.Region.code" telling the region for this row.}
\item{hmw_region_code_col}{The metadata column "HMW.Region.code" telling the region for this row.}
\item{region_code_col}{The metadata column "Region.code" telling the region for this row.}
\item{agg_code_col}{The metadata column "Agg.Code", representing the country, or country group code for individual country level data to be aggregated in to.}
\item{mapping_sheet}{The name of the sheet for name mappings.}
\item{species}{The name of the species column in the concordance spreadsheet.}
}
}
\usage{
conc_cols
}
\description{
A string list containing the names of the columns containing country code information.
}
\examples{
conc_cols
}
\keyword{datasets}
