% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{fill_ilo_data}
\alias{fill_ilo_data}
\title{Fill missing data for the number of hours worked and employed persons based on the earliest value}
\usage{
fill_ilo_data(
  .df,
  country_col = MWTools::conc_cols$country_col,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  sector_col = MWTools::mw_constants$sector_col,
  year = MWTools::mw_constants$year,
  yearly_working_hours_ilo_col = MWTools::ilo_cols$yearly_working_hours_ilo_col,
  employed_persons_ilo_col = MWTools::ilo_cols$employed_persons_ilo_col
)
}
\arguments{
\item{employed_persons_ilo_col}{}
}
\value{

}
\description{
Fill missing data for the number of hours worked and employed persons based on the earliest value
}
