% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hmw_functions.R
\name{get_broad.sector_data}
\alias{get_broad.sector_data}
\title{Retrieve data with sectors organised by "Broad sector"}
\usage{
get_broad.sector_data(.df)
}
\arguments{
\item{.df}{}
}
\value{

}
\description{
Retrieve data with sectors organised by "Broad sector"
}
\examples{

}
