% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_hmw_data_functions.R
\name{get_ilo_hmw_data}
\alias{get_ilo_hmw_data}
\title{Retrieve ILO data for working hours and employed persons}
\usage{
get_ilo_hmw_data(
  country_name = MWTools::mw_constants$country_name,
  sex = MWTools::mw_constants$sex,
  sector = MWTools::mw_constants$sector,
  year = MWTools::mw_constants$year,
  working_hours = MWTools::mw_constants$working_hours,
  employed_persons = MWTools::mw_constants$employed_persons
)
}
\arguments{
\item{country_name, sex, sector, year, working_hours, employed}{persons See \code{MWtools::mw_constants()}.}
}
\value{

}
\description{
Use the \code{Rilostat} package to create a data frame containing mean yearly
working hours and number of employed persons by country, sex, sector, and year.
}
\examples{
ilo_hmw_data <- get_ilo_hmw_data()

}
