% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_hmw_data_functions.R
\name{get_ilo_hmw_data}
\alias{get_ilo_hmw_data}
\title{Retrieve ILO data for working hours and employed persons}
\usage{
get_ilo_hmw_data(
  country_name = MWTools::mw_constants$country_name,
  sex = MWTools::ilo_cols$sex,
  sector = MWTools::ilo_cols$sector,
  year = MWTools::mw_constants$year,
  working_hours = MWTools::ilo_cols$working_hours,
  employed_persons = MWTools::ilo_cols$employed_persons,
  working_hours_code = MWTools::ilo_codes$working_hours_code,
  employment_code = MWTools::ilo_codes$employment_code
)
}
\arguments{
\item{country_name, sex, sector, year, working_hours, employed_persons}{See \code{MWtools::mw_constants()}.}

\item{working_hours_code, employment_code}{See \code{MWTools::ilo_codes}.}
}
\value{

}
\description{
Use the \code{Rilostat} package to create a data frame containing mean yearly
working hours and number of employed persons by country, sex, sector, and year.
}
\examples{
ilo_hmw_data <- get_ilo_hmw_data()

}
