% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{get_working_species}
\alias{get_working_species}
\title{Create a data frame containing the total number of animals of working species}
\usage{
get_working_species(
  .df,
  species = MWTools::mw_constants$species,
  country_name = MWTools::mw_constants$country_name,
  country_code_col = MWTools::conc_cols$country_code_col,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  year = MWTools::mw_constants$year,
  value = MWTools::mw_constants$value,
  asses = MWTools::mw_species$asses,
  camels = MWTools::mw_species$camels,
  cattle = MWTools::mw_species$cattle,
  horses = MWTools::mw_species$horses,
  mules = MWTools::mw_species$mules,
  buffaloes = MWTools::mw_species$buffaloes,
  camelids_other = MWTools::mw_species$camelids_other,
  camelids = MWTools::mw_species$camelids,
  live_animals_col = MWTools::amw_analysis_constants$live_animals_col
)
}
\arguments{
\item{.df}{A tidy data frame containing the number of live animals by
country over time. Usually supplied from the FAO through
the functions \code{MWTools::down_fao_live_animals} and
\code{MWTools::tidy_fao_live_animals}.}

\item{species, value}{See \code{MWTools::mw_constants}.}

\item{asses, camels, cattle, horses, mules, buffaloes, camelids_other, camelids}{See \code{MWTools::mw_species}.}
}
\value{

}
\description{
Using a tidy dataframe containing the number of live animals, this function
restricts the species of animals to: Asses, Camels, Cattle, Horses, Mules,
Buffaloes, and Camelids, other; then combines Camels and Camlelids, other
into a combined "Camelids" species group.
}
\examples{
live_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes() \%>\%
  trim_fao_data() \%>\%
  get_working_species()

}
