% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmw_analysis_constants}
\alias{hmw_analysis_constants}
\title{Human muscle work (hmw) analysis data constants}
\format{
A string list with 1 entries.
\describe{
\item{total_working_hours_ilo_col}{The data column "Total.hours \link{hours/year}" containing the number of hours worked yearly by all humans for a given country, sex, sector, and year.}
}
}
\usage{
hmw_analysis_constants
}
\description{
A string list containing a selection of hmw analysis constants.
}
\examples{
hmw_analysis_constants
}
\keyword{datasets}
