% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ilo_cols}
\alias{ilo_cols}
\title{ILO data column names}
\format{
A string list with 4 entries.
\describe{
\item{sex_ilo_col}{The string "Sex".}
\item{sector_ilo_col}{The string "Sector".}
\item{working_hours_ilo_col}{The string "Working.hours".}
\item{employed_persons_ilo_col}{The string "Employed.persons".}
}
}
\usage{
ilo_cols
}
\description{
A string list containing a selection of ILO data column names. For data retrieved by the package \code{Rilostat}.
}
\examples{
ilo_cols
}
\keyword{datasets}
