% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_constants_mw}
\alias{phi_constants_mw}
\title{Muscle work phi constants}
\format{
A data frame with 6 rows. Each row provides the phi value for one energy carrier:
\describe{
\item{food}{The name for food final energy.}
\item{feed}{The name for feed final energy.}
\item{biomass}{The name for biomass primary energy.}
\item{hu_mech}{The name for human mechanical work useful energy.}
\item{an_mech}{The name for animal mechanical work useful energy.}
\item{an_p}{The name for animal propulsion work useful energy.}
}
}
\usage{
phi_constants_mw
}
\description{
A data frame containing
- muscle work products (energy carriers),
- phi (exergy-to-energy ratio) values,
- and an "is.useful" column saying whether that energy carrier is found
at the useful stage.
}
\examples{
phi_constants_mw
}
\keyword{datasets}
