% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amw_functions.R
\name{plot_amw_summary}
\alias{plot_amw_summary}
\title{Title}
\usage{
plot_amw_summary(
  amw_pfu_df,
  amw_numbers_df,
  country,
  sector,
  year = MWTools::mw_constants$year,
  country_code_col = MWTools::conc_cols$country_code_col,
  species = MWTools::mw_constants$species,
  sector_col = MWTools::mw_constants$sector_col,
  stage_col = MWTools::mw_constants$stage_col,
  energy_mj_year = MWTools::mw_constants$energy_mj_year,
  working_animals_col = MWTools::amw_analysis_constants$working_animals_col,
  live_animals_col = MWTools::amw_analysis_constants$live_animals_col
)
}
\arguments{
\item{live_animals_col}{}
}
\value{

}
\description{
Title
}
