% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_mw_df}
\alias{prep_mw_df}
\title{Prep PFU data for conversion to PSUT matrices}
\usage{
prep_mw_df(
  .hmw_pfu_df,
  .amw_pfu_df,
  biomass = MWTools::mw_products$biomass,
  resources = MWTools::mw_sectors$resources_sector,
  farms = MWTools::mw_sectors$farms,
  product = MWTools::mw_constants$product,
  matnames = MWTools::mw_constants$matnames_col,
  rownames = MWTools::mw_constants$rownames_col,
  colnames = MWTools::mw_constants$colnames_col,
  rowtypes = MWTools::mw_constants$rowtypes_col,
  coltypes = MWTools::mw_constants$coltypes_col,
  R_mat = MWTools::psut_cols$R,
  U_mat = MWTools::psut_cols$U,
  V_mat = MWTools::psut_cols$V,
  Y_mat = MWTools::psut_cols$Y,
  industry_type = MWTools::mw_constants$industry_type,
  product_type = MWTools::mw_constants$product_type,
  product_notation = RCLabels::from_notation,
  resource_notation = RCLabels::of_notation,
  species_notation = RCLabels::arrow_notation
)
}
\arguments{
\item{.hmw_pfu_df}{A PFU data frame for human muscle work,
the output from \code{specify_primary_production()}.}

\item{.amw_pfu_df}{A PFU data frame for animal muscle work,
the output from \code{specify_primary_production()}.}
}
\value{
A data frame with correct rows (many duplicated) and a matrix name column.
}
\description{
Primary-final-useful (PFU) data need to be prepared
for conversion to PSUT matrices by
duplicating some rows and adding a matrix name column.
This function performs those operations.
}
\details{
This function binds \code{.hmw_pfu_df} and \code{.amw_pfu_df} by rows
before starting its work.
}
