% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_psut}
\alias{prep_psut}
\title{Prep PFU data for conversion to PSUT matrices}
\usage{
prep_psut(
  .df,
  biomass = MWTools::mw_products$biomass,
  resources = MWTools::mw_sectors$resources_sector,
  farms = MWTools::mw_sectors$farms,
  product = MWTools::mw_constants$product,
  matnames = MWTools::mat_meta_cols$matnames,
  rownames = MWTools::mat_meta_cols$rownames,
  colnames = MWTools::mat_meta_cols$colnames,
  rowtypes = MWTools::mat_meta_cols$rowtypes,
  coltypes = MWTools::mat_meta_cols$coltypes,
  R_name = MWTools::psut_cols$R,
  U_name = MWTools::psut_cols$U,
  V_name = MWTools::psut_cols$V,
  Y_name = MWTools::psut_cols$Y,
  industry_type = MWTools::row_col_types$industry,
  product_type = MWTools::row_col_types$product,
  product_notation = RCLabels::from_notation,
  resource_notation = RCLabels::of_notation,
  species_notation = RCLabels::arrow_notation
)
}
\arguments{
\item{.df}{A data frame containing muscle work data,
likely the output from \code{specify_last_stages()}.}
}
\value{
A data frame containing PSUT matrices.
}
\description{
Primary-final-useful (PFU) data need to be converted to PSUT matrices.
This function take the output of \code{specify_last_stages()}
and converts to PSUT matrices in a data frame.
}
\examples{
hmw_df <- hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu()
amw_df <- amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu()
specify_product(hmw_df, amw_df) \%>\%
  MWTools::specify_primary_production() \%>\%
  specify_useful_products() \%>\%
  specify_fu_machines() \%>\%
  specify_last_stages() \%>\%
  prep_psut()
}
