% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_primary_production}
\alias{specify_primary_production}
\title{Add primary production to a data frame of PFU muscle work data}
\usage{
specify_primary_production(.df)
}
\arguments{
\item{.df}{A muscle work data frame with products already specified,
usually by \code{specify_product()}.}
}
\value{
A data frame with
}
\description{
Adds rows for biomass from resources
}
