% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_primary_production}
\alias{specify_primary_production}
\title{Add primary production to a data frame of PFU muscle work data}
\usage{
specify_primary_production(
  .df,
  product = IEATools::iea_cols$product,
  primary = IEATools::all_stages$primary,
  stage = MWTools::mw_constants$stage_col,
  notation = RCLabels::from_notation,
  resources = MWTools::mw_sectors$resources_sector
)
}
\arguments{
\item{.df}{A muscle work data frame with products already specified,
usually by \code{specify_product()}.}

\item{product}{See \code{IEATools::iea_cols}.}

\item{primary}{See \code{IEATools::all_stages}.}

\item{stage}{See \code{MWTools::mw_constants}.}

\item{notation}{The notation to be used for primary energy product specification.
Default is \code{IEATools::from_notation}.}

\item{resources}{See \code{MWTools::mw_sectors}.}
}
\value{
A data frame with
}
\description{
Adds rows for biomass from resources
}
\examples{
hmw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_hmw_pfu() \%>\%
  specify_product() \%>\%
  specify_primary_production()
amw_test_data_path() \%>\%
  read.csv() \%>\%
  calc_amw_pfu() \%>\%
  specify_product() \%>\%
  specify_primary_production()
}
