% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_amw_data_functions.R
\name{tidy_fao_live_animals}
\alias{tidy_fao_live_animals}
\title{Tidy live animals data from the FAO}
\usage{
tidy_fao_live_animals(
  data_folder,
  country_name = MWTools::mw_constants$country_name,
  species = MWTools::mw_constants$species,
  year = MWTools::mw_constants$year,
  unit = MWTools::mw_constants$unit,
  value = MWTools::mw_constants$value,
  area_fao_col = MWTools::fao_cols$area_fao_col,
  item_fao_col = MWTools::fao_cols$item_fao_col,
  year_fao_col = MWTools::fao_cols$year_fao_col,
  unit_fao_col = MWTools::fao_cols$unit_fao_col,
  value_fao_col = MWTools::fao_cols$value_fao_col
)
}
\arguments{
\item{data_folder}{A user-specified file path to the directory in which to
download the FAO live animals data.}

\item{country_name, species, year, unit, value}{See \code{MWTools::mw_constants}.}

\item{area_fao_col, item_fao_col, year_fao_col, unit_fao_col, value_fao_col}{See \code{MWTools::fao_cols}.}
}
\value{

}
\description{
Create a tidy data frame using data supplied by the function
\code{MWTools::down_fao_live_animals}.
}
\examples{
tidy_live_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data"))

}
