% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_amw_functions.R
\name{tidy_pfu_data}
\alias{tidy_pfu_data}
\title{Tidy a data frame containg primary, final, and useful energy data for working animals}
\usage{
tidy_pfu_data(
  .df,
  country_code_col = MWTools::conc_cols$country_code_col,
  year = MWTools::mw_constants$year,
  species = MWTools::mw_constants$species,
  sector_col = MWTools::mw_constants$sector_col,
  stage_col = MWTools::mw_constants$stage_col,
  energy_mj_year = MWTools::mw_constants$energy_mj_year,
  mw_region_code_col = MWTools::amw_analysis_constants$mw_region_code_col,
  useful_energy_total = MWTools::amw_analysis_constants$useful_energy_total,
  useful_energy_ag = MWTools::amw_analysis_constants$useful_energy_ag,
  useful_energy_tr = MWTools::amw_analysis_constants$useful_energy_tr,
  final_energy_total = MWTools::amw_analysis_constants$final_energy_total,
  final_energy_ag = MWTools::amw_analysis_constants$final_energy_ag,
  final_energy_tr = MWTools::amw_analysis_constants$final_energy_tr,
  primary_energy_total = MWTools::amw_analysis_constants$primary_energy_total,
  primary_energy_ag = MWTools::amw_analysis_constants$primary_energy_ag,
  primary_energy_tr = MWTools::amw_analysis_constants$primary_energy_tr
)
}
\arguments{
\item{.df}{A data frame containing ...}

\item{country_code_col}{See \code{MWTools::conc_cols}.}

\item{year, species, sector_col, stage_col, energy_mj_year}{See \code{MWTools::mw_constants}.}

\item{mw_region_code_col}{See \code{MWTools::amw_analysis_constants}.}

\item{useful_energy_total, useful_energy_ag, useful_energy_tr}{See \code{MWTools::amw_analysis_constants}.}

\item{final_energy_total, final_energy_ag, final_energy_tr}{See \code{MWTools::amw_analysis_constants}.}

\item{primary_energy_total, primary_energy_ag, primary_energy_tr}{See \code{MWTools::amw_analysis_constants}.}
}
\value{

}
\description{
Tidy a data frame containg primary, final, and useful energy data for working animals
}
\examples{
working_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes() \%>\%
  trim_fao_data() \%>\%
  get_working_species() \%>\%
  calc_working_animals() \%>\%
  calc_sector_split() \%>\%
  calc_yearly_feed() \%>\%
  calc_final_energy() \%>\%
  calc_primary_energy() \%>\%
  calc_useful_energy() \%>\%
  tidy_pfu_data()

}
