% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_amw_data_functions.R
\name{trim_fao_data}
\alias{trim_fao_data}
\title{Trim countries in FAO live animals data}
\usage{
trim_fao_data(
  .df,
  country_incl_col = MWTools::conc_cols$country_incl_col,
  country_code_pfu_col = MWTools::conc_cols$country_code_pfu_col
)
}
\arguments{
\item{.df}{The data frame containing live animals data.}

\item{country_incl_col, country_code_pfu_col}{See \code{MWTools::conc_cols}.}
}
\value{

}
\description{
This function filters out countries which have more than one instance
and aggregate regions from FAO live animals data. Data is usually downloaded
through the \code{FAOSTAT} package, using the \code{MWTools::down_fao_live_animals}
function, and after applying the \code{MWTools::add_concordance_codes} function.
}
\examples{
live_animals_data <- MWTools::down_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  tidy_fao_live_animals(data_folder = file.path(fs::home_path(), "FAO_data")) \%>\%
  add_concordance_codes() \%>\%
  trim_fao_data()

}
