% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_constants}
\alias{unit_constants}
\title{Unit constants}
\format{
A string list with 5 entries.
\describe{
\item{kcal_to_mj}{The value 0.0041858, representing the conversion factor between kilocalories (kcal) and megajoules (MJ).}
\item{hours_to_seconds}{The value 3600, representing the conversion factor between hours and seconds.}
\item{joules_to_megajoules}{The value 0.000001, representing the conversion factor from joules (J) and megajoules (MJ).}
\item{EJ_to_ktoe}{The value 23884.58966275, representing the conversion factor from exajoules (EJ) to kilotonnes of oil equivalent (ktoe).}
\item{EJ_to_TJ}{The value \code{1e6}, representing the conversion factor from exajoules (EJ) to terajoules (TJ).}
}
}
\usage{
unit_constants
}
\description{
A list containing data values for converting between different units of measurement.
}
\examples{
unit_constants
}
\keyword{datasets}
